requirements.txt
bitrix24_rest==2.0.3
pandas==2.2.3
psycopg2_binary==2.9.9
Requests==2.32.3
SQLAlchemy==2.0.35
streamlit==1.36.0
plotly==5.24.1